Fix redundant header writing
For some reason, we end up emitting headers multiple times
which corrupts the stream obviously.

I have no idea why this is happening.  It might have to do
with passing the response writer by value into io.Copy
and somehow it gets reinitialized?

Anyways...that was extremely painful to debug.